+++ CheeseQuest2/game/menu/MenuManager.java
@@ -1,15 +1,17 @@
 package game.menu;
 
+import game.system.*;
 /**
  * Gets input String from InputPanel, prrocesses it, outputs it to game menu
  * Acts as a shell for menu types
  */
-public class MenuManager {
+public class MenuManager extends Outputable {
     public static final int VERB = 0;
     private String input;
     private String output;
     private Menu menu;
-    private Menu lastMenu;
+    private Menu previousMenu;
+    private Menu ghostPreviousMenu;
 
     private static MenuManager instance;
     /**
@@ -49,11 +51,23 @@ public class MenuManager {
     }
 
     /**
-     * Returns value of lastMenu
+     * Returns value of previousMenu
      * @return
      */
-    public Menu getLastMenu() {
-        return this.lastMenu;
+    public Menu getPreviousMenu() {
+        if (this.menu instanceof GhostMenu) {
+            return this.ghostPreviousMenu;
+        } else {
+            return this.previousMenu;
+        }
+    }
+
+    /**
+     * Returns value of ghostPreviousMenu
+     * @return
+     */
+    public Menu getGhostPreviousMenu() {
+        return this.ghostPreviousMenu;
     }
 
     /**
@@ -61,16 +75,32 @@ public class MenuManager {
      * @param
      */
     public void setMenu(Menu menu) {
-        setLastMenu(this.menu);
+        if (this.menu == null) {
+            this.menu = menu;
+        }
+        setPreviousMenu(this.menu,menu);
     	this.menu = menu;
+        // outputln("After set: Current menu is " + this.menu.getClass().getSimpleName());
         this.menu.outputPrompt();
     }
 
     /**
-     * Sets new value of lastMenu
+     * Sets new value of previousMenu
      * @param
      */
-    public void setLastMenu(Menu lastMenu) {
-        this.lastMenu = lastMenu;
+    public void setPreviousMenu(Menu previousMenu, Menu newMenu) {
+        // outputln("Before set: Current menu is " + newMenu.getClass().getSimpleName());
+        if (newMenu instanceof GhostMenu) {
+            this.ghostPreviousMenu = previousMenu;
+            // outputln("which is a GhostMenu, so will return to " + ghostPreviousMenu.getClass().getSimpleName());
+        } else if (!(previousMenu instanceof GhostMenu)) {
+            this.previousMenu = previousMenu;
+        }
+        // outputln("previousMenu is " + this.previousMenu.getClass().getSimpleName());
+        // if (this.previousMenu == null) {
+        //     System.out.println("previousMenu is null");
+        // } else {
+        //     System.out.println("previousMenu is " + this.previousMenu.getClass().getSimpleName());
+        // }
     }
 }