+++ CheeseQuest2/game/system/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Used by objects in game
- */
-/**
- * @author Quan
- *
- */
-package game.system;